:truck: Move session tests to a new file
Separate user and session tests.
