+++ api/tests/tests_utils.go
@@ -37,3 +37,11 @@ func GenerateRandomUser() interfaces.User {
 	user.Username = FakerInstance.Internet().User()
 	return user
 }
+
+// InsertUser inserts a random user in the database. Use this function when you need to test
+// endpoints that require a user to be logged in whintout having to test the signup endpoint
+func InsertUser(user interfaces.User, router *gin.Engine, handler gin.HandlerFunc) {
+	router.POST("/signup", handler)
+	w, req := SetupPostRequest("/signup", user)
+	router.ServeHTTP(w, req)
+}
